requirements.txt
crewai>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
soundfile>=0.12.1
phonemizer>=3.2.1
torch>=2.0.0
torchaudio>=2.0.0
gradio>=4.0.0
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
